configure eslint, esbuild and husky
