+++ package.json
@@ -1,26 +1,72 @@
 {
-  "name": "commonxs",
+  "name": "co2m.js",
   "version": "1.0.0",
   "description": "",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
-  "scripts": {
-    "test": "test:jest",
-    "build": "tsc",
-    "test:jest": "jest --colors --ci --updateSnapshot --coverage"
+  "type": "module",
+  "lint-staged": {
+    "*.[tj]s": [
+      "./node_modules/prettier/bin/prettier.cjs --tab-width 2 --write ./src",
+      "git add"
+    ]
   },
-  "keywords": [],
-  "author": "",
+  "scripts": {
+    "precommit": "lint-staged",
+    "test": "jest --colors --ci --updateSnapshot --coverage",
+    "test:jest": "jest --colors --ci --updateSnapshot --coverage",
+    "jest:clear": "jest --clear cache --colors --ci --notity",
+    "jest:watch": "jest  --colors --ci --coverage --watch",
+    "clean": "rm -rf coverage; rm -rf dist || exit 0",
+    "deleteModules": "rm -rf node_modules",
+    "dev": "node ./tasks/dev.js",
+    "build": "node ./tasks/build.js",
+    "lint": "eslint"
+  },
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js",
+      "require": "./dist/index.cjs"
+    }
+  },
+  "keywords": [
+    "JavaScript",
+    "Common",
+    "JS",
+    "Utilities"
+  ],
+  "homepage": "co2m-js.nivekaa.com",
+  "bugs": {
+    "email": "admin@nivekaa.com",
+    "url": "https://github.com/nivekalara237/co2mjs/issues"
+  },
+  "author": {
+    "name": "Kevin L. Kemta",
+    "url": "nivekaa.com",
+    "email": "admin@nivekaa.com"
+  },
   "license": "ISC",
   "devDependencies": {
+    "@eslint/js": "^9.9.1",
     "@jest/globals": "^29.7.0",
     "@tsconfig/node-lts": "^20.1.3",
     "@types/jest": "^29.5.12",
     "@types/node": "^20.14.9",
+    "esbuild": "^0.23.1",
+    "esbuild-node-externals": "^1.14.0",
+    "eslint": "^9.9.1",
+    "globals": "^15.9.0",
+    "husky": "^9.1.5",
     "jest": "^29.7.0",
+    "jest-config": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
+    "lint-staged": "^15.2.9",
     "node-notifier": "^10.0.1",
+    "npm-commands": "^1.2.1",
+    "prettier": "^3.3.3",
     "ts-jest": "^29.1.5",
-    "typescript": "^5.5.3"
+    "typescript": "^5.5.3",
+    "typescript-eslint": "^8.3.0"
   }
 }
