+++ src/lib/utils/branch.utils.ts
@@ -1,0 +1,9 @@
+export class BranchUtils {
+  public static betweenRangeNumber = (
+    value: number,
+    min: number,
+    max: number,
+  ): boolean => {
+    return value <= max && value >= min;
+  };
+}
