+++ src/lib/utils/numeric.utils.ts
@@ -1,0 +1,5 @@
+export class NumericUtils {
+  public static isNumeric = (val: any) => {
+    return !Array.isArray(val) && val - parseFloat(val) + 1 >= 0;
+  };
+}
