--- src/utils/branch.utils.ts
@@ -1,5 +1,0 @@
-export class BranchUtils {
-    public static betweenRangeNumber = (value: number, min: number, max: number): boolean => {
-        return value <= max && value >= min;
-    };
-}
