--- src/utils/numeric.utils.ts
@@ -1,5 +1,0 @@
-export class NumericUtils {
-    public static isNumeric = (val: any) => {
-        return !Array.isArray(val) && val - parseFloat(val) + 1 >= 0;
-    };
-}
